Fixed behavior of EditHotkeyPrompt
The RecordHotkeyLabel will now stop listening to the keyboard when the
default button is clicked or when another RecordHotkeyLabel is focused.

Previously, the app would crash if another label was clicked. If the
default button was clicked before time ran out, it would still record
key presses until the timer reached 0.